requirements.txt
# Django Core
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.3.0

# CORS & Filtering
django-cors-headers==4.3.1
django-filter==24.2

# Environment Variables (optional but recommended)
python-dotenv==1.0.0

# AI Integration (Secure Proxy Pattern)
google-genai==1.14.0

# Markdown Editor
django-mdeditor==0.1.18

# API Documentation
drf-spectacular==0.28.0

# Production Server
gunicorn==23.0.0
whitenoise==6.8.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1

# Linting
ruff==0.9.7
